cmd/snap: move key=value attribute parsing to commmon
Skills will reuse this part of capability code.
